Frontend library (vminfo_procps) for procps
Develop the new library `vminfo_procps' which make use of the
procps-newlib library (not yet released).
See: https://gitlab.com/procps-ng/procps/tree/newlib

This library is only compiled when the `configure' script is
run with the option `--enable-libprocps' (and the required
libraries are detected).

At this stage of development, `vminfo_procps' must be considered
a poc.

Signed-off-by: Davide Madrisan <davide.madrisan@gmail.com>